requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
wikipedia-api==0.6.0
googlesearch-python==1.2.3
python-socketio==5.11.0
nltk==3.8.1 